backend/requirements.txt
# Agentic RAG Phase 1 - 統計 API 後端
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic>=2.0.0

# Phase 1 PostgreSQL 與 ETL
psycopg2-binary>=2.9.0
pgvector>=0.3.0
